fix(storage): harden barrier accounting ownership
